+++ library/src/main/java/com/prolificinteractive/materialcalendarview/MaterialCalendarView.java
@@ -13,10 +13,13 @@ import android.util.AttributeSet;
 import android.util.Log;
 import android.util.SparseArray;
 import android.util.TypedValue;
+import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.prolificinteractive.materialcalendarview.format.ArrayWeekDayFormatter;
@@ -52,12 +55,12 @@ import java.util.Set;
  * {@linkplain com.prolificinteractive.materialcalendarview.OnDateChangedListener}
  * </p>
  *
- * @see R.styleable#MaterialCalendarView_arrowColor
- * @see R.styleable#MaterialCalendarView_selectionColor
- * @see R.styleable#MaterialCalendarView_headerTextAppearance
- * @see R.styleable#MaterialCalendarView_dateTextAppearance
- * @see R.styleable#MaterialCalendarView_weekDayTextAppearance
- * @see R.styleable#MaterialCalendarView_showOtherDates
+ * @see R.styleable#MaterialCalendarView_mcv_arrowColor
+ * @see R.styleable#MaterialCalendarView_mcv_selectionColor
+ * @see R.styleable#MaterialCalendarView_mcv_headerTextAppearance
+ * @see R.styleable#MaterialCalendarView_mcv_dateTextAppearance
+ * @see R.styleable#MaterialCalendarView_mcv_weekDayTextAppearance
+ * @see R.styleable#MaterialCalendarView_mcv_showOtherDates
  */
 public class MaterialCalendarView extends FrameLayout {
 
@@ -87,10 +90,9 @@ public class MaterialCalendarView extends FrameLayout {
     private final OnClickListener onClickListener = new OnClickListener() {
         @Override
         public void onClick(View v) {
-            int id = v.getId();
-            if(id == R.id.mcv_calendar_widget_button_forward) {
+            if(v == buttonFuture) {
                 pager.setCurrentItem(pager.getCurrentItem() + 1, true);
-            } else if(id == R.id.mcv_calendar_widget_button_backwards) {
+            } else if(v == buttonPast) {
                 pager.setCurrentItem(pager.getCurrentItem() - 1, true);
             }
         }
@@ -115,7 +117,8 @@ public class MaterialCalendarView extends FrameLayout {
 
     private int accentColor = 0;
     private int arrowColor = Color.BLACK;
-    private CalendarDay currentDate;
+
+    private LinearLayout root;
 
     public MaterialCalendarView(Context context) {
         this(context, null);
@@ -127,12 +130,12 @@ public class MaterialCalendarView extends FrameLayout {
         setClipChildren(false);
         setClipToPadding(false);
 
-        LayoutInflater.from(getContext()).inflate(R.layout.mcv_calendar_widget, this);
+        buttonPast = new DirectionButton(getContext());
+        title = new TextView(getContext());
+        buttonFuture = new DirectionButton(getContext());
+        pager = new ViewPager(getContext());
 
-        title = (TextView) findViewById(R.id.mcv_calendar_widget_title);
-        buttonPast = (DirectionButton) findViewById(R.id.mcv_calendar_widget_button_backwards);
-        buttonFuture = (DirectionButton) findViewById(R.id.mcv_calendar_widget_button_forward);
-        pager = (ViewPager) findViewById(R.id.mcv_pager);
+        setupChildren();
 
         title.setOnClickListener(onClickListener);
         buttonPast.setOnClickListener(onClickListener);
@@ -154,6 +157,12 @@ public class MaterialCalendarView extends FrameLayout {
         TypedArray a = context.getTheme()
                 .obtainStyledAttributes(attrs, R.styleable.MaterialCalendarView, 0, 0);
         try {
+
+            int tileSize = a.getDimensionPixelSize(R.styleable.MaterialCalendarView_mcv_tileSize, -1);
+            if(tileSize > 0) {
+                setTileSize(tileSize);
+            }
+
             setArrowColor(a.getColor(
                 R.styleable.MaterialCalendarView_mcv_arrowColor,
                 Color.BLACK
@@ -203,6 +212,50 @@ public class MaterialCalendarView extends FrameLayout {
         setCurrentDate(currentMonth);
     }
 
+    private void setupChildren() {
+        int tileSize = (int) TypedValue.applyDimension(
+                TypedValue.COMPLEX_UNIT_DIP,
+                getResources().getInteger(R.integer.mcv_default_tile_size),
+                getResources().getDisplayMetrics()
+        );
+
+        root = new LinearLayout(getContext());
+        root.setOrientation(LinearLayout.VERTICAL);
+        root.setClipChildren(false);
+        root.setClipToPadding(false);
+        LayoutParams p = new LayoutParams(
+                tileSize * MonthView.DEFAULT_DAYS_IN_WEEK,
+                tileSize * (MonthView.DEFAULT_MONTH_TILE_HEIGHT + 1)
+        );
+        p.gravity = Gravity.CENTER;
+        addView(root, p);
+
+        LinearLayout topbar = new LinearLayout(getContext());
+        topbar.setOrientation(LinearLayout.HORIZONTAL);
+        topbar.setClipChildren(false);
+        topbar.setClipToPadding(false);
+        root.addView(topbar, new LinearLayout.LayoutParams(LayoutParams.MATCH_PARENT, 0, 1));
+
+        buttonPast.setScaleType(ImageView.ScaleType.CENTER_INSIDE);
+        buttonPast.setImageResource(R.drawable.mcv_action_previous);
+        topbar.addView(buttonPast, new LinearLayout.LayoutParams(0, LayoutParams.MATCH_PARENT, 1));
+
+        title.setGravity(Gravity.CENTER);
+        topbar.addView(title, new LinearLayout.LayoutParams(
+                0, LayoutParams.MATCH_PARENT, MonthView.DEFAULT_DAYS_IN_WEEK - 2
+        ));
+
+        buttonFuture.setScaleType(ImageView.ScaleType.CENTER_INSIDE);
+        buttonFuture.setImageResource(R.drawable.mcv_action_next);
+        topbar.addView(buttonFuture, new LinearLayout.LayoutParams(0, LayoutParams.MATCH_PARENT, 1));
+
+        pager.setId(R.id.mcv_pager);
+        pager.setOffscreenPageLimit(1);
+        root.addView(pager, new LinearLayout.LayoutParams(
+                LayoutParams.MATCH_PARENT, 0, MonthView.DEFAULT_MONTH_TILE_HEIGHT
+        ));
+    }
+
     /**
      * Sets the listener to be notified upon selected date change.
      *
@@ -218,6 +271,32 @@ public class MaterialCalendarView extends FrameLayout {
         }
         buttonPast.setEnabled(canGoBack());
         buttonFuture.setEnabled(canGoForward());
+    }
+
+    /**
+     * Set the size of each tile that makes up the calendar.
+     * Each day is 1 tile, so the widget is 7 tiles wide and 8 tiles tall.
+     *
+     * @param size the new size for each tile in pixels
+     */
+    public void setTileSize(int size) {
+        LayoutParams p = new LayoutParams(
+                size * MonthView.DEFAULT_DAYS_IN_WEEK,
+                size * (MonthView.DEFAULT_MONTH_TILE_HEIGHT + 1)
+        );
+        p.gravity = Gravity.CENTER;
+        root.setLayoutParams(p);
+    }
+
+    /**
+     * @see #setTileSize(int)
+     *
+     * @param tileSizeDp the new size for each tile in dips
+     */
+    public void setTileSizeDp(int tileSizeDp) {
+        setTileSize((int) TypedValue.applyDimension(
+                TypedValue.COMPLEX_UNIT_DIP, tileSizeDp, getResources().getDisplayMetrics()
+        ));
     }
 
     /**
@@ -345,7 +424,7 @@ public class MaterialCalendarView extends FrameLayout {
      * @return The current day shown, will be set to first day of the month
      */
     public CalendarDay getCurrentDate() {
-        return currentDate;
+        return adapter.getItem(pager.getCurrentItem());
     }
 
     /**
@@ -700,8 +779,7 @@ public class MaterialCalendarView extends FrameLayout {
         @Override
         public Object instantiateItem(ViewGroup container, int position) {
             CalendarDay month = months.get(position);
-            MonthView monthView =
-                (MonthView) inflater.inflate(R.layout.mcv_month_view, container, false);
+            MonthView monthView = new MonthView(container.getContext());
             monthView.setTag(TAG_ITEM, month);
 
             monthView.setWeekDayFormatter(weekDayFormatter);