+++ library/src/main/java/com/prolificinteractive/materialcalendarview/MonthView.java
@@ -1,9 +1,8 @@
 package com.prolificinteractive.materialcalendarview;
 
 import android.content.Context;
-import android.util.AttributeSet;
 import android.view.View;
-import android.widget.GridLayout;
+import android.widget.LinearLayout;
 
 import com.prolificinteractive.materialcalendarview.format.WeekDayFormatter;
 
@@ -20,11 +19,15 @@ import static java.util.Calendar.SUNDAY;
  * Display a month of {@linkplain DayView}s and
  * seven {@linkplain WeekDayView}s.
  */
-class MonthView extends GridLayout implements View.OnClickListener {
+class MonthView extends LinearLayout implements View.OnClickListener {
 
-    public static interface Callbacks {
+    protected static final int DEFAULT_DAYS_IN_WEEK = 7;
+    protected static final int DEFAULT_MAX_WEEKS = 6;
+    protected static final int DEFAULT_MONTH_TILE_HEIGHT = DEFAULT_MAX_WEEKS + 1;
 
-        public void onDateChanged(CalendarDay date);
+    public interface Callbacks {
+
+        void onDateChanged(CalendarDay date);
     }
 
     private Callbacks callbacks;
@@ -46,40 +49,44 @@ class MonthView extends GridLayout implements View.OnClickListener {
 
 
     public MonthView(Context context) {
-        this(context, null);
-    }
+        super(context);
 
-    public MonthView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-
-        setColumnCount(7);
-        setRowCount(7);
+        setOrientation(VERTICAL);
 
         setClipChildren(false);
         setClipToPadding(false);
-    }
 
-    @Override
-    protected void onFinishInflate() {
-        super.onFinishInflate();
-
-        int children = getChildCount();
-        for(int i = 0; i < children; i++) {
-            View child = getChildAt(i);
-            if(child instanceof WeekDayView) {
-                weekDayViews.add((WeekDayView) child);
-            } else if(child instanceof DayView) {
-                monthDayViews.add((DayView) child);
-                child.setOnClickListener(this);
+        LinearLayout row = makeRow(this);
+        for (int i = 0; i < DEFAULT_DAYS_IN_WEEK; i++) {
+            WeekDayView weekDayView = new WeekDayView(context);
+            weekDayViews.add(weekDayView);
+            row.addView(weekDayView, new LayoutParams(0, LayoutParams.MATCH_PARENT, 1f));
+        }
+        for(int r = 0; r < DEFAULT_MAX_WEEKS; r++) {
+            row = makeRow(this);
+            for(int i = 0; i < DEFAULT_DAYS_IN_WEEK; i++) {
+                DayView dayView = new DayView(context);
+                dayView.setOnClickListener(this);
+                monthDayViews.add(dayView);
+                row.addView(dayView, new LayoutParams(0, LayoutParams.MATCH_PARENT, 1f));
             }
         }
+
         setFirstDayOfWeek(firstDayOfWeek);
         setSelectedDate(new CalendarDay());
     }
 
+
     public void setDayViewDecorators(Collection<DayViewDecorator> dayViewDecorators) {
         this.dayViewDecorators = dayViewDecorators;
         updateUi();
+    }
+
+    private static LinearLayout makeRow(LinearLayout parent) {
+        LinearLayout row = new LinearLayout(parent.getContext());
+        row.setOrientation(HORIZONTAL);
+        parent.addView(row, new LayoutParams(LayoutParams.MATCH_PARENT, 0, 1f));
+        return row;
     }
 
     public void setWeekDayTextAppearance(int taId) {
@@ -116,7 +123,7 @@ class MonthView extends GridLayout implements View.OnClickListener {
         //If the delta is positive, we want to remove a week
         boolean removeRow = showOtherDates ? delta >= 0 : delta > 0;
         if(removeRow) {
-            delta -= 7;
+            delta -= DEFAULT_DAYS_IN_WEEK;
         }
         tempWorkingCalendar.add(DATE, delta);
         return tempWorkingCalendar;