+++ sample/src/main/java/com/prolificinteractive/materialcalendarview/sample/DynamicSettersActivity.java
@@ -1,22 +1,28 @@
 package com.prolificinteractive.materialcalendarview.sample;
 
+import android.app.AlertDialog;
 import android.app.DatePickerDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
 import android.widget.DatePicker;
+import android.widget.NumberPicker;
 import android.widget.TextView;
+
+import com.prolificinteractive.materialcalendarview.CalendarDay;
+import com.prolificinteractive.materialcalendarview.MaterialCalendarView;
+import com.prolificinteractive.materialcalendarview.OnDateChangedListener;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Random;
+
 import butterknife.ButterKnife;
 import butterknife.InjectView;
 import butterknife.OnCheckedChanged;
 import butterknife.OnClick;
-import com.prolificinteractive.materialcalendarview.CalendarDay;
-import com.prolificinteractive.materialcalendarview.MaterialCalendarView;
-import com.prolificinteractive.materialcalendarview.OnDateChangedListener;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Random;
 
 public class DynamicSettersActivity extends ActionBarActivity implements OnDateChangedListener {
 
@@ -24,12 +30,15 @@ public class DynamicSettersActivity extends ActionBarActivity implements OnDateC
 
     @InjectView(R.id.textView) TextView textView;
     @InjectView(R.id.calendarView) MaterialCalendarView widget;
+    private int currentTileSize;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_dynamic_setters);
         ButterKnife.inject(this);
+
+        currentTileSize = getResources().getInteger(R.integer.mcv_default_tile_size);
 
         widget.setOnDateChangedListener(this);
     }
@@ -94,6 +103,24 @@ public class DynamicSettersActivity extends ActionBarActivity implements OnDateC
         });
         widget.setArrowColor(color);
         widget.setSelectionColor(color);
+    }
+
+    @OnClick(R.id.button_set_tile_size) void onTileSizeClicked() {
+        final NumberPicker view = new NumberPicker(this);
+        view.setMinValue(24);
+        view.setMaxValue(64);
+        view.setWrapSelectorWheel(false);
+        view.setValue(currentTileSize);
+        new AlertDialog.Builder(this)
+                .setView(view)
+                .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        currentTileSize = view.getValue();
+                        widget.setTileSizeDp(currentTileSize);
+                    }
+                })
+                .show();
     }
 
     public static void showDatePickerDialog(Context context, CalendarDay day,