+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Change Log
 ==========
 
+Version 0.3.0 *(2015-04-15)*
+----------------------------
+
+ * Change: Namespaced library resources. Everything is now prefixed with `mcv_`
+ * New: Widget resizing is now possible by setting the tileSize
+     * Change: XML layouts are no longer used
+     * Change: Default tile size is now an integer resource
+
 Version 0.2.5 *(2015-03-11)*
 ----------------------------